+++ src/main/java/com/data/pivot/plugin/model/BaseAnAction.java
@@ -1,6 +1,7 @@
 package com.data.pivot.plugin.model;
 
 import com.data.pivot.plugin.config.DataPivotInitializer;
+import com.intellij.openapi.actionSystem.ActionUpdateThread;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.CommonDataKeys;
@@ -21,6 +22,11 @@ public abstract class BaseAnAction extends AnAction {
         psiElement = e.getData(CommonDataKeys.PSI_ELEMENT);
         loadDataGridInfo(e.getProject());
         action(e);
+    }
+
+    @Override
+    public @NotNull ActionUpdateThread getActionUpdateThread() {
+        return ActionUpdateThread.BGT;
     }
 
     protected abstract void action(AnActionEvent e);